Implement advanced tree view with GitHub API calls
