+++ src/components/ui/tree-view/TreeViewBasicPreview.vue
@@ -1,4 +1,5 @@
 <template>
+
   <vuestic-tree-root>
 
     <vuestic-tree-category label="Electronics">
@@ -9,10 +10,10 @@
 
     <vuestic-tree-category isOpen label="Products">
 
-      <vuestic-tree-category label="GitHub repos">
-        <vuestic-tree-node v-for="repo in repos" :key="repo.id">
-          {{ repo.name }}
-        </vuestic-tree-node>
+      <vuestic-tree-category label="Cables">
+        <vuestic-tree-node>Audio</vuestic-tree-node>
+        <vuestic-tree-node>Video</vuestic-tree-node>
+        <vuestic-tree-node>Optical</vuestic-tree-node>
       </vuestic-tree-category>
 
       <vuestic-tree-node>Monitors</vuestic-tree-node>
@@ -27,30 +28,12 @@
     </vuestic-tree-category>
 
   </vuestic-tree-root>
+
 </template>
 
 <script>
-import to from 'await-to-js'
-import { fetchReposFromUser } from '../../../services/api/userService'
 
 export default {
-  name: 'tree-view-basic-preview',
-  data () {
-    return {
-      repos: [],
-    }
-  },
-  methods: {
-    async fetchRepos () {
-      let err, repos;
-      [ err, repos ] = await to(fetchReposFromUser('mathilde-lannes'))
-      // TODO display error notification if there's an error
-      if (err) return
-      this.repos = repos.data
-    }
-  },
-  mounted: function () {
-    this.fetchRepos()
-  }
+  name: 'tree-view-basic-preview'
 }
 </script>
