+++ src/i18n/en.json
@@ -308,7 +308,7 @@
     "icons": "Icons",
     "selectable": "Selectable",
     "editable": "Editable",
-    "advanced": "Advanced"
+    "advanced": "Advanced with API calls"
   },
   "cards": {
     "cards": "Cards",
