+++ src/services/api/userService.js
@@ -1,3 +1,3 @@
 import { API } from '../api'
 
-export const fetchReposFromUser = (username) => API.get(`users/${username}/repos`)
+export const fetchReposFromUser = username => API.get(`users/${username}/repos`)
